// src/common/types/movie.ts
export interface Movie {
  id: number;
  title: string;
  overview: string;
  poster_path: string | null;
  backdrop_path: string | null;
  release_date: string;
  vote_average: number;
  vote_count: number;
  genre_ids: number[];
  adult: boolean;
  original_language: string;
  original_title: string;
  popularity: number;
  video: boolean;
  // Campos adicionales para detalles
  runtime?: number;
  genres?: Genre[];
  budget?: number;
  revenue?: number;
  status?: string;
  tagline?: string;
}

export interface Genre {
  id: number;
  name: string;
  color?: string;
}

export interface MoviesResponse {
  page: number;
  results: Movie[];
  total_pages: number;
  total_results: number;
}

export interface GenresResponse {
  genres: Genre[];
}